Move user follow and openid into models/user
