+++ models/statistic.go
@@ -7,6 +7,7 @@ package models
 import (
 	"code.gitea.io/gitea/models/db"
 	"code.gitea.io/gitea/models/login"
+	user_model "code.gitea.io/gitea/models/user"
 	"code.gitea.io/gitea/models/webhook"
 	"code.gitea.io/gitea/modules/setting"
 )
@@ -92,7 +93,7 @@ func GetStatistic() (stats Statistic) {
 
 	stats.Counter.Comment, _ = e.Count(new(Comment))
 	stats.Counter.Oauth = 0
-	stats.Counter.Follow, _ = e.Count(new(Follow))
+	stats.Counter.Follow, _ = e.Count(new(user_model.Follow))
 	stats.Counter.Mirror, _ = e.Count(new(Mirror))
 	stats.Counter.Release, _ = e.Count(new(Release))
 	stats.Counter.LoginSource = login.CountSources()
