+++ models/user/main_test.go
@@ -15,5 +15,7 @@ func TestMain(m *testing.M) {
 	db.MainTest(m, filepath.Join("..", ".."),
 		"email_address.yml",
 		"user_redirect.yml",
+		"follow.yml",
+		"user_open_id.yml",
 	)
 }
