+++ models/user/openid_test.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-package models
+package user
 
 import (
 	"testing"
@@ -26,25 +26,6 @@ func TestGetUserOpenIDs(t *testing.T) {
 	if assert.NoError(t, err) && assert.Len(t, oids, 1) {
 		assert.Equal(t, "https://domain1.tld/user2/", oids[0].URI)
 		assert.True(t, oids[0].Show)
-	}
-}
-
-func TestGetUserByOpenID(t *testing.T) {
-	assert.NoError(t, db.PrepareTestDatabase())
-
-	_, err := GetUserByOpenID("https://unknown")
-	if assert.Error(t, err) {
-		assert.True(t, IsErrUserNotExist(err))
-	}
-
-	user, err := GetUserByOpenID("https://user1.domain1.tld")
-	if assert.NoError(t, err) {
-		assert.Equal(t, int64(1), user.ID)
-	}
-
-	user, err = GetUserByOpenID("https://domain1.tld/user2/")
-	if assert.NoError(t, err) {
-		assert.Equal(t, int64(2), user.ID)
 	}
 }
